+++ driver_test.go
@@ -1336,7 +1336,16 @@ func TestPodmanDriver_DefaultCaps(t *testing.T) {
 	require.Contains(t, inspectData.EffectiveCaps, "CAP_CHOWN")
 }
 
-// check modified capabilities (CapAdd/CapDrop)
+// check default process label
+func TestPodmanDriver_DefaultProcessLabel(t *testing.T) {
+	taskCfg := newTaskConfig("", busyboxLongRunningCmd)
+	inspectData := startDestroyInspect(t, taskCfg, "defaultprocesslabel")
+
+	// a default container gets "disable" process label
+	require.Contains(t, inspectData.ProcessLabel, "label=disable")
+}
+
+// check modified capabilities (CapAdd/CapDrop/SelinuxOpts)
 func TestPodmanDriver_Caps(t *testing.T) {
 	taskCfg := newTaskConfig("", busyboxLongRunningCmd)
 	// 	cap_add = [
@@ -1347,6 +1356,10 @@ func TestPodmanDriver_Caps(t *testing.T) {
 	//     "MKNOD",
 	//   ]
 	taskCfg.CapDrop = []string{"CHOWN"}
+	// 	selinux_opts = [
+	//     "disable",
+	//   ]
+	taskCfg.SelinuxOpts = []string{"disable"}
 
 	inspectData := startDestroyInspect(t, taskCfg, "caps")
 
@@ -1354,6 +1367,8 @@ func TestPodmanDriver_Caps(t *testing.T) {
 	require.Contains(t, inspectData.EffectiveCaps, "CAP_SYS_TIME")
 	// we dropped CAP_CHOWN, so we should NOT see it in inspect
 	require.NotContains(t, inspectData.EffectiveCaps, "CAP_CHOWN")
+	// we added "disable" process label, so we should see it in inspect
+	require.Contains(t, inspectData.ProcessLabel, "label=disable")
 }
 
 // check enabled tty option