+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## UNRELEASED
+## 0.4.1 (November 15, 2022)
 
 FEATURES: